+++ src/main/java/com/example/demo/repository/DrugRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Drug;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DrugRepository extends JpaRepository<Drug, Long> {
+public interface DrugRepository extends JpaRepository<Drug, Integer> {
 }
